ut5/educa/requirements.txt
django
pillow
django-braces
django-embed-video
pymemcache
django-debug-toolbar
redis
django-redisboard
djangorestframework
requests
